+++ app/components/paper-select.js
@@ -1,0 +1,2 @@
+import PaperSelect from 'ember-paper/components/paper-select';
+export default PaperSelect;