+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - [#138](https://github.com/miguelcobain/ember-paper/pull/138) Use sidenav and backdrop angular material styles. Use [ember-css-transitions](https://github.com/peec/ember-css-transitions).
 - [#168](https://github.com/miguelcobain/ember-paper/pull/168) Added dynamic height to textarea
 - [#176](https://github.com/miguelcobain/ember-paper/pull/176) Documentation fix for autocomplete
+- [#170](https://github.com/miguelcobain/ember-paper/pull/170) Pass through input helper attributes
 
 ### 0.2.7 (Aug 11, 2015)
 - [#132](https://github.com/miguelcobain/ember-paper/pull/132) Added autocomplete component.