+++ CHANGELOG.md
@@ -1,10 +1,17 @@
 # Ember Paper Changelog
 
+### 0.2.9 (Sep 20, 2015)
+- [#140](https://github.com/miguelcobain/ember-paper/pull/140) Implement Material Menu and Select component.
+
+
 ### 0.2.8 (Aug 19, 2015)
 - [#154](https://github.com/miguelcobain/ember-paper/pull/154) Add support for inline paper-icon in paper-input component
 - [#152](https://github.com/miguelcobain/ember-paper/pull/152) Add support for .md-actions to {{paper-card}}
 - [#151](https://github.com/miguelcobain/ember-paper/pull/151) Only show validation errors if user has interacted with the element
 - [#138](https://github.com/miguelcobain/ember-paper/pull/138) Use sidenav and backdrop angular material styles. Use [ember-css-transitions](https://github.com/peec/ember-css-transitions).
+- [#168](https://github.com/miguelcobain/ember-paper/pull/168) Added dynamic height to textarea
+- [#176](https://github.com/miguelcobain/ember-paper/pull/176) Documentation fix for autocomplete
+- [#170](https://github.com/miguelcobain/ember-paper/pull/170) Pass through input helper attributes
 
 ### 0.2.7 (Aug 11, 2015)
 - [#132](https://github.com/miguelcobain/ember-paper/pull/132) Added autocomplete component.